shifted api call for image id into the SpellDescImg component, added useEffect to only call once
