+++ src/spellStub.js
@@ -1,32 +1,11 @@
 import React, { useState } from "react";
 import PrepareBtn from "./PrepareBtn";
 import SpellDescImg from "./SpellDescImg";
-import axios from "axios";
+// import axios from "axios";
 
 export default function SpellStub(props) {
-  const [dropdown, SetDropdown] = useState(false);
-  const [imageUrl, SetImageUrl] = useState({});
-
-  function DropDownSpellDesc() {
-    SetDropdown(!dropdown);
-
-    function handleResponse(response) {
-      let drive_id = response.data[0].drive_id;
-      let spellUrl = `https://drive.google.com/uc?export=view&id=${drive_id}`;
-      let url_dict = {};
-      url_dict["image_url"] = spellUrl;
-
-      SetImageUrl(url_dict);
-      console.log(spellUrl);
-    }
-
-    let spell_name = props.spell_name.toLowerCase();
-    let apiUrl = `/image_descs/${spell_name}`;
-    axios
-      .get(apiUrl)
-      .then(handleResponse)
-      .catch((err) => console.log(err));
-  }
+  const [dropdown, SetDropDown] = useState(false);
+  const showImage = () => SetDropDown(!dropdown);
 
   return (
     <div className="SpellStub">
@@ -35,7 +14,7 @@ export default function SpellStub(props) {
           <span
             className="spell-display d-flex"
             id="spell_name"
-            onClick={DropDownSpellDesc}
+            onClick={showImage}
           >
             {props.spell_name}
           </span>
@@ -51,7 +30,7 @@ export default function SpellStub(props) {
             : "desc-dropdown desc-dropdown-hidden d-none"
         }
       >
-        <SpellDescImg imageURL={imageUrl["image_url"]} />
+        <SpellDescImg name={props.spell_name} dropdown={dropdown} />
       </div>
     </div>
   );
